changed card update to a streamed process to decrease memory usage (prevent OOM exception...)

+++ lib/util/fs_data.dart
@@ -26,8 +26,10 @@ Future<int> fsGetImagesSize() async {
 }
 
 Future<void> deleteDatabaseData() async {
+  await MTGDB.closeDB();
   final File dbFile = File(join(await getDatabasesPath(), 'mtg_moe_database.db'));
-  dbFile.deleteSync();
+  if (dbFile.existsSync())
+    dbFile.deleteSync();
   MTGDB.invalidate();
 }
